src/requirements.txt
psycopg2-binary
pandas
streamlit
plotly
matplotlib
langchain
langchain-community
langchain-openai
langgraph
openai
python-dotenv
polars
